refactor(directory): :recycle: styles/responsive/ -> utils/responsive/, styles/components/ -> components/common/
