+++ app/src/components/common/Avatar.tsx
