+++ app/src/components/common/Center.tsx
@@ -1,4 +1,3 @@
-import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
 type CenterProps = Partial<{
@@ -6,14 +5,10 @@ type CenterProps = Partial<{
   h: string;
 }>;
 
-export const center = css`
+export const Center = styled.div<CenterProps>`
   display: flex;
   justify-content: center;
   align-items: center;
-`;
-
-export const Center = styled.div<CenterProps>`
-  ${center}
   width: ${({ w = 'auto' }) => w};
   height: ${({ h = 'auto' }) => h};
 `;
