+++ app/src/components/common/Text.tsx
