+++ app/src/components/common/index.ts
@@ -1,3 +1,4 @@
+/* Chakra UI like components */
 export * from './Center';
 export * from './Stack';
 export * from './Text';
@@ -5,3 +6,10 @@ export * from './Divider';
 export * from './Spinner';
 export * from './Spacer';
 export * from './Avatar';
+
+/* Component that extends HTML element */
+export * from './Button';
+export * from './Input';
+
+/* etc */
+export * from './Overlay';
