+++ app/src/components/elements/Header/Desktop/index.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { UserSearchBar } from '@/components/elements/UserSearchBar';
-import { HStack } from '@/styles/components';
+import { HStack } from '@/components/common';
 
 export const DesktopHeader = () => {
   return (
